moving to imagestack module
